# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: API for managing group food orders.
  title: GroupGrub
  version: 1.0.0
servers:
  - description: Production server
    url: https://api.groupgrub.com
paths:
...
  /participants:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Participant'
                type: array
          description: Successfully retrieved participants.
        "404":
          description: No participants found.
      summary: Retrieve all participants registered in the service.
      x-accepts:
        - application/json
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Participant'
        required: true
      responses:
        "201":
          description: Participants added successfully.
        "400":
          description: Failed to add participants.
      summary: Add new participants.
      x-content-type: application/json
      x-accepts:
        - application/json
  /participants/{participantID}:
    delete:
      parameters:
        - description: The unique identifier of the participant.
          explode: false
          in: path
          name: participantID
          required: true
          schema:
            type: string
          style: simple
      responses:
        "204":
          description: Participant deleted successfully.
        "404":
          description: Participant not found.
      summary: Delete a participant.
      x-accepts:
        - application/json
    get:
      parameters:
        - description: The unique identifier of the participant.
          explode: false
          in: path
          name: participantID
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
          description: Participant details retrieved successfully.
        "404":
          description: Participant not found.
      summary: Retrieve details of a participant by participantID.
      x-accepts:
        - application/json
    put:
      parameters:
        - description: The unique identifier of the participant.
          explode: false
          in: path
          name: participantID
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Participant'
        required: true
      responses:
        "200":
          description: Participant updated successfully.
        "400":
          description: Failed to update participant.
        "404":
          description: Participant not found.
      summary: Update a participant's details.
      x-content-type: application/json
      x-accepts:
        - application/json
  /participants/{participantID}/orders/{participantOrderID}:
    delete:
      parameters:
        - description: The unique identifier of the participant.
          explode: false
          in: path
          name: participantID
          required: true
          schema:
            type: string
          style: simple
        - description: The unique identifier of the participant's order.
          explode: false
          in: path
          name: participantOrderID
          required: true
          schema:
            type: string
          style: simple
      responses:
        "204":
          description: Order deleted successfully.
        "404":
          description: Order not found.
      summary: Delete a participant's order.
      x-accepts:
        - application/json
    get:
      parameters:
        - description: The unique identifier of the participant.
          explode: false
          in: path
          name: participantID
          required: true
          schema:
            type: string
          style: simple
        - description: The unique identifier of the participant's order.
          explode: false
          in: path
          name: participantOrderID
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipantOrder'
          description: Participant's order retrieved successfully.
        "404":
          description: Order not found.
      summary: Retrieve a specific participant's order.
      x-accepts:
        - application/json
    put:
      parameters:
        - description: The unique identifier of the participant.
          explode: false
          in: path
          name: participantID
          required: true
          schema:
            type: string
          style: simple
        - description: The unique identifier of the participant's order.
          explode: false
          in: path
          name: participantOrderID
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipantOrder'
        required: true
      responses:
        "200":
          description: Order updated successfully.
        "400":
          description: Failed to update order.
        "404":
          description: Order not found.
      summary: Update a participant's order.
      x-content-type: application/json
      x-accepts:
        - application/json
  /participants/{participantID}/orders:
    get:
      parameters:
        - description: The unique identifier of the participant.
          explode: false
          in: path
          name: participantID
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ParticipantOrder'
                type: array
          description: List of participant orders retrieved successfully.
      summary: Retrieve all participant orders for a specific participant.
      x-accepts:
        - application/json
    post:
      parameters:
        - description: The unique identifier of the participant.
          explode: false
          in: path
          name: participantID
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipantOrder'
        required: true
      responses:
        "201":
          description: Participant order created successfully.
      summary: Create a new participant order.
      x-content-type: application/json
      x-accepts:
        - application/json
components:
  schemas:
...
    Participant:
      example:
        participantID: participantID
        participantOrderIDs:
          - participantOrderIDs
          - participantOrderIDs
        name: name
      properties:
        participantID:
          type: string
        name:
          type: string
        participantOrderIDs:
          items:
            type: string
          type: array
      type: object
    ParticipantOrder:
      example:
        menuItemIDs:
          key: 6
        comments: comments
        participantOrderID: participantOrderID
      properties:
        participantOrderID:
          type: string
        menuItemIDs:
          additionalProperties:
            type: integer
          type: object
        comments:
          type: string
      type: object
...